put helpers in folder, add libsnark flags
